Add files via upload
This is a guided project I worked on, on Coursera on the fundamentals of python. While I was working on this course I learned a lot about python programming. I learned about loops, functions, variables and variable assignments, how to do Mathematical calculations in python, strings in python, and lists in python. As I worked on this project. I got to independently create a guessing game in python. I independently wrote a code for generating a random number using a random number generator in python between a particular range, letting the user enter a number they guess from 1 to 100, and then use a while loop to produce outputs about the outcome of the guessing game.