Add files via upload
This was a hands-on-lab in which I learned how to create data visualization in Excel as part of the IBM Data Analyst Professional Certificate on Coursera. I used Pivot tables, slicers, and timelines to manipulate data and show how data manipulation in pivot tables in Excel translate into changes in the pivot charts that are based on pivot tables used to model data, and how those changes affect data visualizations in Excel.